fix: allow moving files across drives
